requirements.txt
streamlit
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-readers-wikipedia
wikipedia
llama-index-vector-stores-chroma
chromadb
